refactor(profile): extract shared profile components for E3-03 reuse
Avant E3-03 / E3-05 / E3-06, on extrait les briques visuelles du profil
pour éviter que les stagiaires dupliquent ProfileHeader / ProfileStats /
ProfileTabs / ProfileGridItem / EmptyState / Skeleton.

- Crée src/features/profile/components/ : 7 fichiers UI réutilisables
- Ajoute useUserProfile(userId) dans useProfile.ts pour les profils
  autres que l'user connecté (cache key ['profile', 'user', userId])
- ProfileScreen refactoré : ~250 lignes au lieu de ~650, plus de bricks
  copiables par les écrans à venir
